Rejoice: - Deprecating Universal Renderer, we will just use client.jsx and server.jsx - react-router is now working.
