+++ README.md
@@ -130,10 +130,6 @@ Read more on DO config [here](https://github.com/choonkending/react-webpack-node
 			- Logout.react
 			- About.react
 
-## UniversalRenderer
-
-This is a modified version of alt's IsomorphicRenderer. I wished to use webpack to build my server and client side code, but wanted to easily bootstrap data into the stores, and render the correct component using react-router. This takes into account the fact that we're using a singleton store and flushing it everytime (as opposed to creating an instance everytime).
-
 ## Questions:
 ### 1. Google Authentication does not work locally or on heroku!
