--- app/app.js
@@ -1,5 +1,0 @@
-import universalRenderer from './utils/UniversalRenderer';
-import alt from 'altInstance';
-import html from 'base.html';
-
-export default universalRenderer(alt, html);
