+++ app/server.jsx
@@ -1,0 +1,36 @@
+import Iso from 'iso';
+import React from 'react';
+import Router from 'react-router';
+import Location from 'react-router/lib/Location';
+
+import alt from 'altInstance';
+import routes from 'routes.jsx';
+import html from 'base.html';
+
+/*
+ * @param {AltObject} an instance of the Alt object
+ * @param {ReactObject} routes specified in react-router
+ * @param {Object} Data to bootstrap our altStores with
+ * @param {Object} req passed from Express/Koa server
+ */
+const renderToMarkup = (alt, state, req) => {
+  let markup;
+  let location = new Location(req.path, req.query);
+  Router.run(routes, location, (error, initialState, transition) => {
+    alt.bootstrap(state);
+    let content = React.renderToString(<Router {...initialState} />);
+    markup = Iso.render(content, alt.flush());
+  });
+
+  return markup;
+};
+
+/* 
+ * Export render function to be used in server/config/routes.js
+ * We grab the state passed in from the server and the req object from Express/Koa
+ * and pass it into the Router.run function.
+ */
+export default function render(state, req) {
+  const markup = renderToMarkup(alt, state, req);
+  return html.replace('CONTENT', markup);
+};
