+++ webpack/webpack-hot-dev-server.config.js
@@ -50,7 +50,7 @@ module.exports = [
     entry: {
       app:[ 'webpack-dev-server/client?http://' + WEBPACK_HOST + ":" + WEBPACK_PORT,
        'webpack/hot/only-dev-server',
-        "./app" ]
+        "./client" ]
     },
     output: {
       // The output directory as absolute path
@@ -63,7 +63,7 @@ module.exports = [
     },
     module: {
       preLoaders: [{
-        test: /\.js$|.jsx$/,
+        test: /\.js$|\.jsx$/,
         exclude: /node_modules/,
         loaders: ["eslint"]
       }],
@@ -90,7 +90,7 @@ module.exports = [
     name: "server-side rendering",
     context: path.join(__dirname, "..", "app"),
     entry: {
-      app: "./app",
+      app: "./server",
       header: "./elements/Header"
     },
     target: "node",
